[fix][req]日志表数据复制bugfix

See merge request scm-pms/scm-wip-backstage!54

+++ wip-infrastructure/src/main/java/com/cvte/scm/wip/infrastructure/requirement/mapper/dataobject/WipReqLogDO.java
@@ -50,11 +50,11 @@ public class WipReqLogDO {
 
     @Column(name = "before_item_qty")
     @ApiModelProperty(value = "修改前物料数量")
-    private Integer beforeItemQty;
+    private Long beforeItemQty;
 
     @Column(name = "after_item_qty")
     @ApiModelProperty(value = "修改后物料数量")
-    private Integer afterItemQty;
+    private Long afterItemQty;
 
     @Column(name = "before_wkp_no")
     @ApiModelProperty(value = "修改前工序号")
